official/普通的身份认证器/src/requirements.txt
fastapi==0.61.0
uvicorn==0.11.8
PyJWT==1.5.0  # intentionally here!
cryptography==3.1.1
aiofiles==0.5.0
pyOpenSSL==19.1.0  # dynamic flag check